rmv unnecessary int in month iwr

+++ utils.py
@@ -34,10 +34,10 @@ def writenewIWR(scenario, all_split_data, all_data, firstline_iwr, sow, users,
             row_data[2] = str(int(float(row_data[2]) * remaining_demand))
             # scale other months
             for j in range(len(all_split_data[i + firstline_iwr]) - 2):
-                row_data.append(str(int(float(all_split_data[i + firstline_iwr][j + 1]) * remaining_demand)))
+                row_data.append(str(float(all_split_data[i + firstline_iwr][j + 1]) * remaining_demand))
         else:
             for j in range(len(all_split_data[i + firstline_iwr]) - 2):
-                row_data.append(str(int(float(all_split_data[i + firstline_iwr][j + 1]))))
+                row_data.append(str(float(all_split_data[i + firstline_iwr][j + 1])))
         # append row of adjusted data
         new_data.append(row_data)
